commitlog_replayer: Add exception type for truncation case and throw+handle
Fixes #6653

If a file is either externally truncated or somehow not preallocated fully
(not odsync?), such that when we replay we hit a buffer underflow in entry
read, we should signal this wrapped, instead of just underlying out_of_range,
then handle this as a recoverable/ignorable (but warned) exception in replayer,
same as data corruption, not terminate server.
Yes, there is possible data loss, but again, same as file corruption...

Unit test included.