utils: lister: reimplement rmdir
Calling remove_file for each entry is slow
since it requires to do a full path lookup for each entry.

Instead, this patch introduces lister's own implementation
of synchronous list_directory (allowing yielding
in between buffers), and then uses `unlinkat` to
unlink each file, reusing the open directory
file handle.

Also, the implementation ignores ENOENT errors
as they are expected when rmdir is called in parallel
on the same directory tree.  It makes sense to tolerate it
since the desired effect is to remove the whole tree
anyway.

TODO: provide infrastructure in the seastar library
for unlinkat.

Fixes #10484

Signed-off-by: Benny Halevy <bhalevy@scylladb.com>