loading_cache: add method for adding entries synchronously
In a coming patch, we will want to use the loading_cache
to store udf instances to reuse them. Because the used
instances cannot be used by multiple users simultaneously,
we will want to remove the instance from the cache before
each use, and reinsert the instance to the cache afterwards.
We want to insert the value immediately, so that no other
loader function is called. We cannot achieve this using
loading_cache's 'get_ptr' method, because there is no
guarantee that the future returning the ready value gets
executed before any other future loading the value.

To fix this, we add a method that synchronously adds
a ready value to the cache, without waiting for any
futures. The value is inserted to the cache if there
was no ready value in the cache.
the cache, without waiting for any loader function.

Signed-off-by: Wojciech Mitros <wojciech.mitros@scylladb.com>