database: truncate: skip flush when taking snapshot
database::truncate already flushes the table
on auto_snapshot so there is never a reason
to flush it again in table::snapshot.

Note that cf.can_flush() is false only if memtables
are empty so there nothing to flush or there is
is no seal_immediate_fn and then table::snapshot
wouldn't be able to flush either.

Signed-off-by: Benny Halevy <bhalevy@scylladb.com>