api: storage_service: increase visibility of snapshot ops in the log
snapshot operations over the api are rare
but they contain significant state on disk in the
form of sstables hard-linked to the snapshot directories.

Also, we've seen snapshot operations hang in the field,
requiring a core dump to analyse the issue,
while there were no records in the log indicating
when previous snapshot operations were last executed.

This change promotes logging to info level
when take_snapshot and del_snapshot start,
and logs errors if in case they fail.

Signed-off-by: Benny Halevy <bhalevy@scylladb.com>