utils/logalloc: make tracker a sharded service
The tracker already quacks and walks like a sharded service. It is
constructed on all shards, it has to be configured, and it has a stop
method too. This patch makes this explicit: it is instantiated in main
and in tests as a sharded service or a simple instance where that is
enough.
The formerly separate free function prime_segment_pool() is folded into
tracker::configure(), which in turn is made private and is now called
from the tracker constructor, which gets a tracker::config instance
parameter. To streamline instantiating the tracker, tracker::config
gets a constructor with optional params defaulted to their de-facto
defaults.
Background reclaiming is opt-in, only enabled if there is a
background reclaiming scheduling group passed to the configuration. This
is to allow tests that either can't or won't deal with stopping
background reclaim (which is unnecessary for them).

With all this, the global shard tracker instance is gone and everybody
has to use a local instance.