test/cql-pytest: add test for default clustering order of SELECT
Our documentation for SELECT,
https://docs.scylladb.com/getting-started/dml/#ordering-results
says that:

  "The ORDER BY clause lets you select the order of the returned results.
   It takes as argument a list of column names along with the order for
   the column (ASC for ascendant and DESC for descendant,
   **omitting the order being equivalent to ASC**)."

The test in this patch confirms that the last emphasized line is not
accurate - The default order for SELECT is the default order of the table
being read - NOT always ascending order. If the table was created with
descending WITH CLUSTERING ORDER BY, then a SELECT not specifying
an ORDER BY will get this descending order by default.

The test passes on both Scylla and Cassandra, demonstrating that this
behavior is expected and correct - regardless of what our docs say.

Signed-off-by: Nadav Har'El <nyh@scylladb.com>
Message-Id: <20220515115030.775813-1-nyh@scylladb.com>